deleted extra code.  Added code for modal. Format
