+++ src/index.html
@@ -6,7 +6,6 @@
     <title>Sleep Outside | Home</title>
     <link rel="stylesheet" href="/css/style.css" />
     <script src="js/main.js" type="module"></script>
-    <script src="js/modal.js" type="module"></script>
   </head>
 
   <body>
