--- src/js/modal.js
@@ -1,85 +1,0 @@
-import { calcNumCartItems } from "./cartContents.js";
-import ProductData from "./ProductData.mjs";
-import { productDetailsTemplate } from "./ProductDetails.mjs";
-import { setLocalStorage } from "./utils.mjs";
-
-// create instance of Product Data class
-const mdataSource = new ProductData("tents");
-
-const modal = document.getElementById("modal");
-const buttons = document.getElementsByClassName("card__productModalbtn");
-const closeX = document.getElementsByClassName("close")[0];
-
-// wrap in window onload to remove error
-window.onload = function () {
-  // opens the modal window
-  function openModal(buttonId) {
-    modal.style.display = "block";
-    const product = new ProductModal(buttonId, mdataSource);
-    product.init();
-  }
-
-  // add listeners to the "Quick View" buttons
-  for (var i = 0; i < buttons.length; i++) {
-    let buttonId = buttons[i].value;
-    buttons[i].addEventListener(
-      "click",
-      function () {
-        openModal(buttonId);
-      },
-      false
-    );
-  }
-};
-
-// close modal if click on "x"
-closeX.onclick = function () {
-  removeModalDetails();
-};
-
-// close modal if click outside of it
-window.onclick = function (event) {
-  if (event.target == modal) {
-    removeModalDetails();
-  }
-};
-
-class ProductModal {
-  constructor(productId, dataSource) {
-    this.productId = productId;
-    this.product = {};
-    this.dataSource = dataSource;
-  }
-
-  async init() {
-    // use our datasource to get the details for the current product.
-    this.product = await this.dataSource.findProductById(this.productId);
-    // once we have the product details we can render out the HTML
-    this.renderProductModal("modal-content");
-    document
-      .getElementById("addToCart")
-      .addEventListener("click", this.addToCart.bind(this));
-  }
-
-  // add to cart from home page
-  addToCart() {
-    setLocalStorage("so-cart", this.product);
-    // TEST
-    calcNumCartItems();
-  }
-
-  renderProductModal(selector) {
-    const element = document.getElementsByClassName(selector)[0];
-    element.insertAdjacentHTML(
-      "beforeEnd",
-      productDetailsTemplate(this.product)
-    );
-  }
-}
-
-// remove modal details for clean slate each click
-function removeModalDetails() {
-  modal.style.display = "none";
-  const element = document.getElementsByClassName("product-detail")[0];
-  element.remove();
-}
