+++ src/product_listing/index.html
@@ -5,7 +5,6 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Sleep Outside | Home</title>
     <link rel="stylesheet" href="../css/style.css" />
-    <script src="/js/modal.js" type="module"></script>
     <script src="/js/product_listing.js" type="module"></script>
   </head>
   <body>
